Steps 1 and half of Step 2
Started by dissecting the example and understanding the logic and use of ECS.
Once I was happy I started by adding a new JSON reader and JSON file for wave data. this would dictate the amount of each enemy to spawn each wave.
Then I worked on rewriting the enemy spawner engine logic to follow the json rules and explicitly spawn each mob type x amount of times, following their spawn timer rules
Using local variables in the engine, I started to track the number of mobs killed vs how many for that wave to indicate when the wave is completed and to load the next one.
To get the information to the HUD, I needed to create a Wave Data component which is part of the HUD entity view component, this is used with an implementor that communicates with the UI and shows the number of units left and the wave number on screen.